+++ src/quacc/recipes/onetep/_base.py
@@ -130,7 +130,7 @@ def _prep_calculator(
         pseudo_path=str(SETTINGS.ONETEP_PP_PATH) if SETTINGS.ONETEP_PP_PATH else ".",
         parallel_info=SETTINGS.ONETEP_PARALLEL_CMD,
         profile=OnetepProfile(
-            str(SETTINGS.ONETEP_CMD)
+            SETTINGS.ONETEP_CMD
         ),  # TODO: If the ASE merge is successful, we need to change ONETEP_PARALLEL_CMD to a list[str] and remove parallel info.
         # If we also have access to post_args we can point not to the binary but to the launcher which takes -t nthreads as a post_args
         **calc_flags,